Add failing test case to make sure a resource's context does not change during load/save.
New issue
Summary: Resource's context changes during load/save
A ResourceProxy's .all() or .get_by() will use the context given at query time for the resulting resources. This is not correct when the context is a merge of several contexts as it is with the unamed graph in most (all?) SPARQL implementations.
The solution would be to also query for the resource's context.


git-svn-id: http://surfrdf.googlecode.com/svn/trunk@329 19dffd64-4534-11de-9eaf-578f35f84a2a